Split tests up into separate files
